refactor: move clear-sort button management into TableHeaderRenderer
